Restructures App's conditional rendering
